Only register one click listener
This is the fix for a bug I had on mobile devices, where I needed to register a lot of swipebox instances on a single Page. The swipebox would start to open slower and slower because there were many click listerners on $(document) at the same time. Alternatively if a single click listeners is not wanted, please consider removing the listerner on destroy.